Improve stability. Fix error when user can't add new bookmark because there are no schedule ids. Refactoring

+++ app/src/main/java/com/subreax/schedule/data/repository/bookmark/offline/OfflineBookmarkRepository.kt
@@ -4,8 +4,9 @@ import com.subreax.schedule.data.local.dao.BookmarkDao
 import com.subreax.schedule.data.local.entitiy.BookmarkEntity
 import com.subreax.schedule.data.local.entitiy.asExternalModel
 import com.subreax.schedule.data.model.ScheduleBookmark
-import com.subreax.schedule.data.model.ScheduleId
+import com.subreax.schedule.data.model.ScheduleType
 import com.subreax.schedule.data.repository.bookmark.BookmarkRepository
+import com.subreax.schedule.data.repository.schedule_id.ScheduleIdRepository
 import com.subreax.schedule.utils.Resource
 import com.subreax.schedule.utils.UiText
 import kotlinx.coroutines.CoroutineScope
@@ -20,7 +21,8 @@ import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 class OfflineBookmarkRepository @Inject constructor(
-    private val bookmarkDao: BookmarkDao
+    private val bookmarkDao: BookmarkDao,
+    private val scheduleIdRepository: ScheduleIdRepository
 ) : BookmarkRepository {
     private val coroutineScope = CoroutineScope(SupervisorJob() + Dispatchers.IO)
 
@@ -33,22 +35,37 @@ class OfflineBookmarkRepository @Inject constructor(
     override val bookmarks: Flow<List<ScheduleBookmark>>
         get() = _bookmarks
 
-    override suspend fun addBookmark(scheduleId: ScheduleId): Resource<ScheduleBookmark> {
+    override suspend fun addBookmark(scheduleId: String): Resource<ScheduleBookmark> {
         return coroutineScope.async {
-            // todo: handle errors
-            // todo: check for unique
-            bookmarkDao.addBookmark(scheduleId.asNewBookmarkEntity())
-            val bookmark = bookmarkDao.findBookmark(scheduleId.value)!!.asExternalModel()
-            Resource.Success(bookmark)
+            if (bookmarkDao.isBookmarkExist(scheduleId)) {
+                return@async Resource.Failure(UiText.hardcoded("Закладка уже существует"))
+            }
+
+            val typeRes = getScheduleType(scheduleId)
+            if (typeRes is Resource.Failure) {
+                return@async Resource.Failure(typeRes.message)
+            }
+
+            val bookmark = BookmarkEntity(
+                id = 0,
+                scheduleId = scheduleId,
+                type = typeRes.requireValue(),
+                name = ScheduleBookmark.NO_NAME
+            )
+            bookmarkDao.addBookmark(bookmark)
+            Resource.Success(bookmarkDao.findBookmark(scheduleId)!!.asExternalModel())
         }.await()
     }
 
     override suspend fun deleteBookmark(scheduleId: String): Resource<Unit> {
-        return withContext(Dispatchers.IO) {
-            // todo: check if bookmark exist
-            bookmarkDao.deleteBookmark(scheduleId)
-            Resource.Success(Unit)
-        }
+        return coroutineScope.async {
+            if (bookmarkDao.isBookmarkExist(scheduleId)) {
+                bookmarkDao.deleteBookmark(scheduleId)
+                Resource.Success(Unit)
+            } else {
+                Resource.Failure(UiText.hardcoded("Закладка не существует"))
+            }
+        }.await()
     }
 
     override suspend fun getBookmark(scheduleId: String): Resource<ScheduleBookmark> {
@@ -62,38 +79,26 @@ class OfflineBookmarkRepository @Inject constructor(
         }
     }
 
-    override suspend fun isEmpty(): Boolean {
+    override suspend fun isNotEmpty(): Boolean {
         return withContext(Dispatchers.IO) {
-            bookmarkDao.countBookmarks() == 0
+            bookmarkDao.isNotEmpty()
         }
     }
 
     override suspend fun setBookmarkName(scheduleId: String, name: String): Resource<Unit> {
-        // todo: handle errors
-        return withContext(Dispatchers.IO) {
-            bookmarkDao.updateBookmarkName(scheduleId, name)
-            Resource.Success(Unit)
+        return coroutineScope.async {
+            if (bookmarkDao.isBookmarkExist(scheduleId)) {
+                bookmarkDao.updateBookmarkName(scheduleId, name)
+                Resource.Success(Unit)
+            } else {
+                Resource.Failure(UiText.hardcoded("Закладка не существует"))
+            }
+        }.await()
+    }
+
+    private suspend fun getScheduleType(scheduleId: String): Resource<ScheduleType> {
+        return scheduleIdRepository.getScheduleId(scheduleId).ifSuccess {
+            Resource.Success(it.type)
         }
     }
-
-    private fun ScheduleId.asNewBookmarkEntity(
-        name: String = ScheduleBookmark.NO_NAME
-    ): BookmarkEntity {
-        return BookmarkEntity(
-            id = 0,
-            scheduleId = value,
-            typeValue = type.ordinal,
-            name = name
-        )
-    }
-
-    /*private fun BookmarkEntity.asExternalModel(): ScheduleBookmark {
-        return ScheduleBookmark(
-            scheduleId = ScheduleId(
-                value = scheduleId,
-                type = ScheduleType.entries[typeValue]
-            ),
-            name = name
-        )
-    }*/
 }
